refactor: create reusable flag layout component
