+++ src/app/components/crypto-assets/crypto-currency-asset/crypto-currency-asset-item.layout.tsx
@@ -1,12 +1,13 @@
-import { Box, Stack, StackProps } from '@stacks/ui';
+import { Flex, StackProps } from '@stacks/ui';
 import { forwardRefWithAs } from '@stacks/ui-core';
 
-import type { Money } from '@shared/models/money.model';
+import { Money } from '@shared/models/money.model';
 
 import { getFormattedBalance } from '@app/common/crypto-assets/stacks-crypto-asset.utils';
 import { ftDecimals } from '@app/common/stacks-utils';
 import { usePressable } from '@app/components/item-hover';
-import { SpaceBetween } from '@app/components/space-between';
+import { Flag } from '@app/components/layout/flag';
+import { SpaceBetween } from '@app/components/layout/space-between';
 import { Tooltip } from '@app/components/tooltip';
 import { Caption, Text } from '@app/components/typography';
 
@@ -32,40 +33,26 @@ export const CryptoCurrencyAssetItemLayout = forwardRefWithAs(
     const isUnanchored = !!(subBalance && !balance.amount.isEqualTo(subBalance.amount));
 
     return (
-      <Box
-        as={isPressable ? 'button' : 'div'}
-        display="flex"
-        textAlign="left"
-        outline={0}
-        position="relative"
-        ref={ref}
-        flexGrow={1}
-        spacing="base"
-        {...rest}
-        {...bind}
-      >
-        <Stack alignItems="center" flexGrow={1} isInline spacing="base">
-          {icon}
-          <Stack flexGrow={1} spacing="none">
-            <SpaceBetween width="100%">
-              <Text>{title}</Text>
-              <Tooltip
-                label={formattedBalance.isAbbreviated ? balance.amount.toString() : undefined}
-                placement="left-start"
-              >
-                <Text data-testid={title} fontVariantNumeric="tabular-nums" textAlign="right">
-                  {formattedBalance.value}
-                </Text>
-              </Tooltip>
-            </SpaceBetween>
-            <SpaceBetween height="1.25rem" width="100%">
-              <Caption>{caption}</Caption>
-              {isUnanchored && subBalance ? <SubBalance balance={subBalance} /> : null}
-            </SpaceBetween>
-          </Stack>
-        </Stack>
+      <Flex as={isPressable ? 'button' : 'div'} outline={0} ref={ref} {...rest} {...bind}>
+        <Flag img={icon} align="middle" width="100%">
+          <SpaceBetween width="100%">
+            <Text>{title}</Text>
+            <Tooltip
+              label={formattedBalance.isAbbreviated ? balance.amount.toString() : undefined}
+              placement="left-start"
+            >
+              <Text data-testid={title} fontVariantNumeric="tabular-nums" textAlign="right">
+                {formattedBalance.value}
+              </Text>
+            </Tooltip>
+          </SpaceBetween>
+          <SpaceBetween height="1.25rem" width="100%">
+            <Caption>{caption}</Caption>
+            {isUnanchored && subBalance ? <SubBalance balance={subBalance} /> : null}
+          </SpaceBetween>
+        </Flag>
         {component}
-      </Box>
+      </Flex>
     );
   }
 );
