+++ src/app/components/fees-row/components/fees-row.layout.tsx
@@ -1,7 +1,7 @@
 import { Stack, StackProps } from '@stacks/ui';
 import { useField } from 'formik';
 
-import { SpaceBetween } from '@app/components/space-between';
+import { SpaceBetween } from '@app/components/layout/space-between';
 import { SponsoredLabel } from '@app/components/sponsored-label';
 import { Caption } from '@app/components/typography';
 import { WarningLabel } from '@app/components/warning-label';
