+++ src/app/features/settings-dropdown/components/advanced-menu-items.tsx
@@ -9,7 +9,7 @@ import {
 } from '@shared/logger-storage';
 import { isNumber } from '@shared/utils';
 
-import { Divider } from '@app/components/divider';
+import { Divider } from '@app/components/layout/divider';
 import { Caption } from '@app/components/typography';
 
 import { SettingsMenuItem as MenuItem } from './settings-menu-item';
