+++ src/app/features/settings-dropdown/components/ledger-item-row.tsx
@@ -1,7 +1,7 @@
 import { Flex, Text, color } from '@stacks/ui';
 
-import { Divider } from '@app/components/divider';
 import { WalletTypeLedgerIcon } from '@app/components/icons/wallet-type-ledger-icon';
+import { Divider } from '@app/components/layout/divider';
 
 interface LedgerDeviceItemRowProps {
   deviceType?: string;
