+++ src/app/pages/send/send-crypto-asset-form/components/confirmation/components/confirmation-detail.tsx
@@ -2,7 +2,7 @@ import { ReactNode } from 'react';
 
 import { Text } from '@stacks/ui';
 
-import { SpaceBetween } from '@app/components/space-between';
+import { SpaceBetween } from '@app/components/layout/space-between';
 import { Caption } from '@app/components/typography';
 
 interface ConfirmationDetailProps {
