+++ src/app/pages/transaction-request/components/contract-deploy-details/contract-deploy-details.tsx
@@ -4,7 +4,7 @@ import { BoxProps, CodeBlock, Stack, color } from '@stacks/ui';
 
 import { Prism } from '@app/common/clarity-prism';
 import { useWallet } from '@app/common/hooks/use-wallet';
-import { Divider } from '@app/components/divider';
+import { Divider } from '@app/components/layout/divider';
 import { Caption, Title } from '@app/components/typography';
 import { AttachmentRow } from '@app/pages/transaction-request/components/attachment-row';
 import { ContractPreviewLayout } from '@app/pages/transaction-request/components/contract-preview';
