+++ src/app/pages/transaction-request/components/stx-transfer-details/stx-transfer-details.tsx
@@ -1,6 +1,6 @@
 import { Stack, color } from '@stacks/ui';
 
-import { Divider } from '@app/components/divider';
+import { Divider } from '@app/components/layout/divider';
 import { Title } from '@app/components/typography';
 import { AttachmentRow } from '@app/pages/transaction-request/components/attachment-row';
 import { Row } from '@app/pages/transaction-request/components/row';
