+++ src/app/pages/transaction-request/components/transaction-error/error-messages.tsx
@@ -11,9 +11,9 @@ import { useAnalytics } from '@app/common/hooks/analytics/use-analytics';
 import { useDrawers } from '@app/common/hooks/use-drawers';
 import { useScrollLock } from '@app/common/hooks/use-scroll-lock';
 import { stacksValue } from '@app/common/stacks-utils';
+import { SpaceBetween } from '@app/components/layout/space-between';
 import { PrimaryButton } from '@app/components/primary-button';
 import { SecondaryButton } from '@app/components/secondary-button';
-import { SpaceBetween } from '@app/components/space-between';
 import { Caption } from '@app/components/typography';
 import { ErrorMessage } from '@app/pages/transaction-request/components/transaction-error/error-message';
 import { useCurrentStacksAccountAnchoredBalances } from '@app/query/stacks/balance/balance.hooks';
